Use Integer rather than Int for Energy, which gets very large

+++ src/Update.hs
@@ -38,7 +38,7 @@ performCommand (botId, cmd) state@State{..} =
           guard $ isValidCoord matrix newBotCoord
           -- TODO: check all coords along path are empty
           pure state { bots = Map.insert botId movedBot bots
-                     , energy = energy + Energy energyToMoveBot
+                     , energy = energy + Energy (fromIntegral energyToMoveBot)
                      }
           where movedBot = bot { coord = newBotCoord }
                 newBotCoord = translateBy vector $ coord bot
@@ -46,7 +46,7 @@ performCommand (botId, cmd) state@State{..} =
 
         LMove (SLD vector1) (SLD vector2) ->
           pure state { bots = Map.insert botId movedBot bots
-                     , energy = energy + Energy energyToMoveBot
+                     , energy = energy + Energy (fromIntegral energyToMoveBot)
                      }
           where movedBot = bot { coord = newBotCoord }
                 newBotCoord = (translateBy vector1 . translateBy vector2) $ coord bot
